+++ src/app/filter-search/filter-search-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { FilterSearchPage } from './filter-search.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: FilterSearchPage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class FilterSearchPageRoutingModule {}
