+++ src/app/filter-search/filter-search.page.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-filter-search',
+  templateUrl: './filter-search.page.html',
+  styleUrls: ['./filter-search.page.scss'],
+})
+export class FilterSearchPage implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
