+++ src/app/tabs/tabs-routing.module.ts
@@ -12,8 +12,8 @@ const routes: Routes = [
         loadChildren: () => import('../tab1/tab1.module').then(m => m.Tab1PageModule)
       },
       {
-        path: 'tab2',
-        loadChildren: () => import('../tab2/tab2.module').then(m => m.Tab2PageModule)
+        path: 'filter-search',
+        loadChildren: () => import('../filter-search/filter-search.module').then(m => m.FilterSearchPageModule)
       },
       {
         path: 'cocktails',
