+++ src/app/tabs/tabs.page.html
@@ -2,19 +2,19 @@
 
   <ion-tab-bar slot="bottom">
     <ion-tab-button tab="tab1" href="/tabs/tab1">
-      <ion-icon aria-hidden="true" name="home"></ion-icon>
+      <ion-icon src="assets/icon/tabsIcons/home.svg" style="fill: black;"></ion-icon>
       <ion-label>Home</ion-label>
     </ion-tab-button>
 
-    <ion-tab-button tab="tab2" href="/tabs/tab2">
-      <ion-icon aria-hidden="true" name="filter-circle"></ion-icon>
-      <ion-label>Filter</ion-label>
+    <ion-tab-button tab="filter-search" href="/tabs/filter-search">
+      <ion-icon src="assets/icon/tabsIcons/search.svg" style="fill: black;"></ion-icon>
+      <ion-label>Search</ion-label>
     </ion-tab-button>
 
     <ion-tab-button tab="cocktails" href="/tabs/cocktails">
-      <ion-icon aria-hidden="true" name="beer"></ion-icon>
+      <ion-icon src="assets/icon/tabsIcons/cocktail.svg"></ion-icon>
       <ion-label>Cocktails</ion-label>
     </ion-tab-button>
+    
   </ion-tab-bar>
-
 </ion-tabs>
